feat: typing stubs of cocotb dut's
I was not able to find a way to prevent having to include tests/stubs/
directory in upstream, even though they are automatically generated.

There is a bit of a chicken and the egg situation, where the module
importing the stubs is required before the stubs actually can be
generated.

Long-term there should be a solution to this, but short-term we include
stubs upstream.